update sysndb
update sysndb

+++ syncdb.py
@@ -26,8 +26,11 @@ class sync():
         self.tables=self.__helper.showsql("show tables")
     def CreateDB(self):
         list=[]
-        for db in self.dbs:
-            list.append(db["Database"])
+        try:
+            for db in self.dbs:
+                list.append(db["Database"])
+        except Exception,e:
+            pass
         if self.database not in list:
             createdbsql = """create database if not exists %s CHARACTER SET utf8 COLLATE utf8_bin""" %(self.database)
             self.__helper.installsql(createdbsql)
@@ -39,9 +42,12 @@ class sync():
         self.tables=self.__helper.showsql("show tables",dbname)
         #获取当前数据库中的表组成列表
         list=[]
-        for tables in self.tables:
-            for table in tables.values():
-                list.append(table)
+        try:
+            for tables in self.tables:
+                for table in tables.values():
+                    list.append(table)
+        except Exception,e:
+            pass
         return list
     def CreteTable(self,dbname):
         tablelist=self.AllTable(dbname)
@@ -58,7 +64,7 @@ class sync():
                             createtablesql="""create table if not exists %s.%s_%s(temp bigint)"""%(self.database,jsontable["name"],i)
                             self.__helper.installsql(createtablesql,dbname)
     def Diff(self,json,tablename,dbname):
-        #当前数据库的数据 
+        #当前数据库的数据
         showtable="""SHOW FULL COLUMNS FROM %s.%s"""%(dbname,tablename)
         temp=self.__helper.showsql(showtable, dbname)
         jsonlist=[]
@@ -76,7 +82,7 @@ class sync():
                     if dbfieldlist == listfield:
                         #如果相等，那么就对比字典是否相同，如果不同，则更新，如果相同，则不做处理
                         for listfieldall in json["fields"]:
-                            for dbfieldall in temp:      
+                            for dbfieldall in temp:
                                 try:
                                     del dbfieldall["Privileges"]
                                     del dbfieldall["Key"]
@@ -87,7 +93,7 @@ class sync():
                                         if dbfieldall !=  listfieldall:
                                             self.Update(dbname, tablename,listfieldall)
                                             self.UpdateFiled(dbname, tablename, listfieldall)
-                                
+
         else:
             temp=set(jsonlist)-set(dblist)
             for tempfile in temp:
@@ -114,7 +120,7 @@ class sync():
         try:
             showtable="""SHOW FULL COLUMNS FROM %s.%s"""%(dbname,tablename)
             temp=self.__helper.showsql(showtable, dbname)
-            for delfield in temp: 
+            for delfield in temp:
                 if delfield["Field"] == "temp":
                     altertablesql="""alter table %s.%s drop column temp"""%(dbname,tablename)
                     self.__helper.installsql(altertablesql,dbname)
@@ -123,15 +129,16 @@ class sync():
     def Index(self,json,tablename,dbname):
         for indexk,indexv in json.items():
             if indexk=="indexs":
-                for vluer in indexv:
-                    try:
-                        delindex="""alter table {0}.{1} drop index {2}""".format(dbname,tablename,vluer["name"])
-                        self.__helper.installsql(delindex, dbname)
-                    except Exception,e:
-                        pass
-                    key=','.join(x for x in vluer["fields"])
-                    indexsql="""alter table {0}.{1}  add index {2}({3})""".format(dbname,tablename,vluer["name"],key)
-                    self.__helper.installsql(indexsql, dbname)
+                if len(json["indexs"]) != 0:
+                    for vluer in indexv:
+                        try:
+                            delindex="""alter table {0}.{1} drop index {2}""".format(dbname,tablename,vluer["name"])
+                            self.__helper.installsql(delindex, dbname)
+                        except Exception,e:
+                            pass
+                        key=','.join(x for x in vluer["fields"])
+                        indexsql="""alter table {0}.{1}  add index {2}({3})""".format(dbname,tablename,vluer["name"],key)
+                        self.__helper.installsql(indexsql, dbname)
     def Update(self,dbname,tablename,field):
         if field["Null"] == "NO":
             if field["Default"] != "null":
@@ -190,7 +197,7 @@ class sync():
     def DiffTable(self,dbname):
         #当前数据库的表
         tablelist=self.AllTable(dbname)
-        #json中文件的表       
+        #json中文件的表
         for jsontable in self.installinfo["tables"]:
             for table in tablelist:
                 if jsontable["shard"] == "1":
@@ -222,7 +229,7 @@ class sync():
                         for temp in json["initDatas"]:
                             fields = ','.join(['`{0}`'.format(x['field']) for x in temp])
                             values = ','.join(['"{0}"'.format(x['value']) for x in temp])
-                            sql = 'INSERT INTO `{0}`.`{1}`({2})VALUES({3})'.format(dbname, tablename, fields, values)        
+                            sql = 'INSERT INTO `{0}`.`{1}`({2})VALUES({3})'.format(dbname, tablename, fields, values)
                             try:
                                 self.__helper.installsql(sql,dbname)
                             except Exception,e:
@@ -258,7 +265,7 @@ class table():
             self.installinfo=json.loads(configinfo)
         self.__helper=sqlhelper.MySqlHelper()
     def AppTables(self):
-        
+
         tables=self.installinfo["tables"]
         for table in tables:
             for version in table["serviceVersion"]:
@@ -266,7 +273,7 @@ class table():
             sql="""insert into IM_DBCONFIG_APPTABLE(appServerID,appServerVersion,tableName) values('{0}','{1}','{2}')""".format(table["service"],sv,table["name"])
             self.__helper.installsql(sql,'IM_DBCONFIG')
     def DateSource(self):
-        
+
         try:
             remark=self.installinfo["remark"]
             database=self.installinfo["database"]
@@ -276,7 +283,7 @@ class table():
             sql="""insert into IM_DBCONFIG_DATASOURCE(dataSourceID,remark) values('{0}','{1}')""".format(database," ")
         self.__helper.installsql(sql,'IM_DBCONFIG')
     def DateSourceSub(self):
-        
+
         try:
             database=self.installinfo["database"]
             ip=self.__helper.mysql_str["host"]
@@ -287,7 +294,7 @@ class table():
         except Exception,e:
             print e
     def Tableds(self):
-        
+
         tables=self.installinfo["tables"]
         database=self.installinfo["database"]
         for table in tables:
@@ -306,7 +313,7 @@ class table():
                     apptableid="""select appTableID from IM_DBCONFIG_APPTABLE where appServerID= '{0}'and tableName='{1}' and appServerVersion='{2}'""".format(table["service"],table["name"],sv)
                     id=self.__helper.select(apptableid,'IM_DBCONFIG')
                     sql="insert into IM_DBCONFIG_TABLEDS(appTableID,tableSegName,dataSourceID) values({0},'{1}','{2}')".format(id["appTableID"],tablename,database)
-                    self.__helper.installsql(sql,'IM_DBCONFIG')      
+                    self.__helper.installsql(sql,'IM_DBCONFIG')
 
 if __name__ ==  "__main__":
     files=os.listdir("/data/im/database/dbsql/")
@@ -314,7 +321,7 @@ if __name__ ==  "__main__":
     for file in files:
             print file
             t=sync(file)
-            t.CreateDB()      
+            t.CreateDB()
             t.CreteTable(t.database)
             t.DiffTable(t.database)
     for file in files:
@@ -322,11 +329,11 @@ if __name__ ==  "__main__":
             t=table(file)
             t.AppTables()
             t.DateSource()
-            t.DateSourceSub() 
+            t.DateSourceSub()
             t.Tableds()
 '''
 t=sync("/data/database/dbsql/IM_ENTERPRISE.json")
-t.CreateDB()      
+t.CreateDB()
 t.CreteTable(t.database)
-t.DiffTable(t.database)  
-'''   
+t.DiffTable(t.database)
+'''
